7_semantic_chunker.py
from langchain_experimental.text_splitter import SemanticChunker
from langchain_openai import OpenAIEmbeddings
from dotenv import load_dotenv

load_dotenv()  # Load environment variables from .env file

# Same Tesla text but structured to show semantic grouping
tesla_text = """Tesla's Q3 Results
Tesla reported record revenue of $25.2B in Q3 2024.
The company exceeded analyst expectations by 15%.
Revenue growth was driven by strong vehicle deliveries.

Model Y Performance  
The Model Y became the best-selling vehicle globally, with 350,000 units sold.
Customer satisfaction ratings reached an all-time high of 96%.
Model Y now represents 60% of Tesla's total vehicle sales.

Production Challenges
Supply chain issues caused a 12% increase in production costs.
Tesla is working to diversify its supplier base.
New manufacturing techniques are being implemented to reduce costs."""

semantic_splitter = SemanticChunker(
    embeddings=OpenAIEmbeddings(),
    # Breakpoint threshold type and amount
    # Options: "percentile", "standard_deviation"
    # "percentile" means the threshold is a percentile of the embedding distances
    # "standard_deviation" means the threshold is a multiple of the standard deviation of the embedding distances
    breakpoint_threshold_type="percentile", 
    # Set the threshold amount (e.g., 70th percentile)
    breakpoint_threshold_amount=70
)

# Split the text into chunks
chunks = semantic_splitter.split_text(tesla_text)

print("SEMANTIC CHUNKING RESULTS:")
print("=" * 50)
for i, chunk in enumerate(chunks, 1):
    print(f"Chunk {i}: ({len(chunk)} chars)")
    print(f'"{chunk}"')
    print()