requirements.txt
langchain-community
langchain-text-splitters
langchain-openai
langchain-chroma
python-dotenv
langchain-experimental